backend/requirements.txt
instaloader
flask
pillow
typing-extensions
python-dotenv
google-generativeai
supabase
icalendar
pytz
gunicorn
